Added Routes and redirects. Added not working ChatPage
